requirements.txt
elevenlabs==0.2.26
python-dotenv==1.0.0
sounddevice==0.4.6
numpy>=1.26.0
scipy>=1.11.0
openai-whisper
pydub 